[fix][node] modify code for scale or remove node

+++ dingoops/api/node.py
@@ -14,14 +14,14 @@ node_service = NodeService()
 
 @router.get("/node/list", summary="k8s集群节点列表", description="k8s集群节点列表")
 async def list_nodes(cluster_id:str = Query(None, description="集群id"),
-        cluster_name:str = Query(None, description="集群名称"),
-        type:str = Query(None, description="节点类型"),
-        page: int = Query(1, description="页码"),
-        page_size: int = Query(10, description="页数量大小"),
-        sort_keys:str = Query(None, description="排序字段"),
-        sort_dirs:str = Query(None, description="排序方式"),):
+                     cluster_name:str = Query(None, description="集群名称"),
+                     type:str = Query(None, description="节点类型"),
+                     page: int = Query(1, description="页码"),
+                     page_size: int = Query(10, description="页数量大小"),
+                     sort_keys:str = Query(None, description="排序字段"),
+                     sort_dirs:str = Query(None, description="排序方式"),):
     try:
-         # 声明查询条件的dict
+        # 声明查询条件的dict
         query_params = {}
         # 查询条件组装
         if cluster_name:
@@ -59,6 +59,8 @@ async def create_node(cluster: ClusterObject):
         # 先检查下是否有正在处于扩容的状态，如果是就直接返回
         cluster_service = ClusterService()
         result = cluster_service.get_cluster(cluster.id)
+        if not result:
+            raise HTTPException(status_code=400, detail="the cluster does not exist, please check")
         if result.status == "scaling":
             raise HTTPException(status_code=400, detail="the cluster is scaling, please wait")
 
@@ -78,8 +80,10 @@ async def delete_node(node_list_info: NodeRemoveObject):
         # 先检查下是否有正在处于缩容的状态，如果是就直接返回
         cluster_service = ClusterService()
         result = cluster_service.get_cluster(node_list_info.cluster_id)
+        if not result:
+            raise HTTPException(status_code=400, detail="the cluster does not exist, please check")
         if result.status == "removing":
-            raise HTTPException(status_code=400, detail="the cluster is scaling, please wait")
+            raise HTTPException(status_code=400, detail="the cluster is removing, please wait")
 
         # 缩容某些节点
         result = node_service.delete_node(node_list_info)
